+++ client/rpc/dex_client.go
@@ -3,18 +3,35 @@ package rpc
 import (
 	"errors"
 	"fmt"
+
+	"github.com/binance-chain/go-sdk/common"
 	"github.com/binance-chain/go-sdk/common/types"
+	"github.com/binance-chain/go-sdk/keys"
+	gtypes "github.com/binance-chain/go-sdk/types"
+	"github.com/binance-chain/go-sdk/types/msg"
 	"github.com/binance-chain/go-sdk/types/tx"
+	core_types "github.com/tendermint/tendermint/rpc/core/types"
+)
+
+type SyncType int
+
+const (
+	Async SyncType = iota
+	Sync
+	Commit
 )
 
 const (
-	AccountStoreName = "acc"
-	TokenStoreName   = "tokens"
-	ParamABCIPrefix  = "param"
+	AccountStoreName    = "acc"
+	ParamABCIPrefix     = "param"
+	TimeLockMsgRoute    = "timelock"
+	AtomicSwapStoreName = "atomic_swap"
+
+	TimeLockrcNotFoundErrorCode = 458760
 )
 
 type DexClient interface {
-	TxInfoSearch(query string, prove bool, page, perPage int) ([]tx.Info, error)
+	TxInfoSearch(query string, prove bool, page, perPage int) ([]Info, error)
 	ListAllTokens(offset int, limit int) ([]types.Token, error)
 	GetTokenInfo(symbol string) (*types.Token, error)
 	GetAccount(addr types.AccAddress) (acc types.Account, err error)
@@ -25,12 +42,28 @@ type DexClient interface {
 	GetFee() ([]types.FeeParam, error)
 	GetOpenOrders(addr types.AccAddress, pair string) ([]types.OpenOrder, error)
 	GetTradingPairs(offset int, limit int) ([]types.TradingPair, error)
-	GetDepth(tradePair string) (*types.OrderBook, error)
+	GetDepth(tradePair string, level int) (*types.OrderBook, error)
 	GetProposals(status types.ProposalStatus, numLatest int64) ([]types.Proposal, error)
 	GetProposal(proposalId int64) (types.Proposal, error)
+	GetTimelocks(addr types.AccAddress) ([]types.TimeLockRecord, error)
+	GetTimelock(addr types.AccAddress, recordID int64) (*types.TimeLockRecord, error)
+	GetSwapByID(swapID types.SwapBytes) (types.AtomicSwap, error)
+	GetSwapByCreator(creatorAddr string, offset int64, limit int64) ([]types.SwapBytes, error)
+	GetSwapByRecipient(recipientAddr string, offset int64, limit int64) ([]types.SwapBytes, error)
+
+	SetKeyManager(k keys.KeyManager)
+	SendToken(transfers []msg.Transfer, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error)
+	CreateOrder(baseAssetSymbol, quoteAssetSymbol string, op int8, price, quantity int64, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error)
+	CancelOrder(baseAssetSymbol, quoteAssetSymbol, refId string, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error)
+	HTLT(recipient types.AccAddress, recipientOtherChain, senderOtherChain string, randomNumberHash []byte, timestamp int64,
+		amount types.Coins, expectedIncome string, heightSpan int64, crossChain bool, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error)
+	DepositHTLT(recipient types.AccAddress, swapID []byte, amount types.Coins,
+		syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error)
+	ClaimHTLT(swapID []byte, randomNumber []byte, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error)
+	RefundHTLT(swapID []byte, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error)
 }
 
-func (c *HTTP) TxInfoSearch(query string, prove bool, page, perPage int) ([]tx.Info, error) {
+func (c *HTTP) TxInfoSearch(query string, prove bool, page, perPage int) ([]Info, error) {
 	if err := ValidateTxSearchQueryStr(query); err != nil {
 		return nil, err
 	}
@@ -49,6 +82,9 @@ func (c *HTTP) ListAllTokens(offset int, limit int) ([]types.Token, error) {
 	if err != nil {
 		return nil, err
 	}
+	if !result.Response.IsOK(){
+		return nil, fmt.Errorf(result.Response.Log)
+	}
 	bz := result.Response.GetValue()
 	tokens := make([]types.Token, 0)
 	err = c.cdc.UnmarshalBinaryLengthPrefixed(bz, &tokens)
@@ -63,6 +99,9 @@ func (c *HTTP) GetTokenInfo(symbol string) (*types.Token, error) {
 	result, err := c.ABCIQuery(path, nil)
 	if err != nil {
 		return nil, err
+	}
+	if !result.Response.IsOK(){
+		return nil, fmt.Errorf(result.Response.Log)
 	}
 	bz := result.Response.GetValue()
 	token := new(types.Token)
@@ -180,6 +219,9 @@ func (c *HTTP) GetFee() ([]types.FeeParam, error) {
 	if err != nil {
 		return nil, err
 	}
+	if !rawFee.Response.IsOK(){
+		return nil, fmt.Errorf(rawFee.Response.Log)
+	}
 	var fees []types.FeeParam
 	err = c.cdc.UnmarshalBinaryLengthPrefixed(rawFee.Response.GetValue(), &fees)
 	return fees, err
@@ -192,6 +234,9 @@ func (c *HTTP) GetOpenOrders(addr types.AccAddress, pair string) ([]types.OpenOr
 	rawOrders, err := c.ABCIQuery(fmt.Sprintf("dex/openorders/%s/%s", pair, addr), nil)
 	if err != nil {
 		return nil, err
+	}
+	if !rawOrders.Response.IsOK(){
+		return nil, fmt.Errorf(rawOrders.Response.Log)
 	}
 	bz := rawOrders.Response.GetValue()
 	openOrders := make([]types.OpenOrder, 0)
@@ -216,6 +261,9 @@ func (c *HTTP) GetTradingPairs(offset int, limit int) ([]types.TradingPair, erro
 	if err != nil {
 		return nil, err
 	}
+	if !rawTradePairs.Response.IsOK(){
+		return nil, fmt.Errorf(rawTradePairs.Response.Log)
+	}
 	pairs := make([]types.TradingPair, 0)
 	if rawTradePairs.Response.GetValue() == nil {
 		return pairs, nil
@@ -224,13 +272,19 @@ func (c *HTTP) GetTradingPairs(offset int, limit int) ([]types.TradingPair, erro
 	return pairs, err
 }
 
-func (c *HTTP) GetDepth(tradePair string) (*types.OrderBook, error) {
+func (c *HTTP) GetDepth(tradePair string, level int) (*types.OrderBook, error) {
 	if err := ValidatePair(tradePair); err != nil {
 		return nil, err
 	}
-	rawDepth, err := c.ABCIQuery(fmt.Sprintf("dex/orderbook/%s", tradePair), nil)
+	if err := ValidateDepthLevel(level); err != nil {
+		return nil, err
+	}
+	rawDepth, err := c.ABCIQuery(fmt.Sprintf("dex/orderbook/%s/%d", tradePair, level), nil)
 	if err != nil {
 		return nil, err
+	}
+	if !rawDepth.Response.IsOK(){
+		return nil, fmt.Errorf(rawDepth.Response.Log)
 	}
 	var ob types.OrderBook
 	err = c.cdc.UnmarshalBinaryLengthPrefixed(rawDepth.Response.GetValue(), &ob)
@@ -238,6 +292,71 @@ func (c *HTTP) GetDepth(tradePair string) (*types.OrderBook, error) {
 		return nil, err
 	}
 	return &ob, nil
+}
+
+func (c *HTTP) GetTimelocks(addr types.AccAddress) ([]types.TimeLockRecord, error) {
+
+	params := types.QueryTimeLocksParams{
+		Account: addr,
+	}
+
+	bz, err := c.cdc.MarshalJSON(params)
+
+	if err != nil {
+		fmt.Errorf("marshal params failed %v", err)
+	}
+
+	rawRecords, err := c.ABCIQuery(fmt.Sprintf("custom/%s/%s", TimeLockMsgRoute, "timelocks"), bz)
+
+	if err != nil {
+		return nil, err
+	}
+	if rawRecords == nil {
+		return nil, fmt.Errorf("zero records")
+	}
+	if !rawRecords.Response.IsOK(){
+		return nil, fmt.Errorf(rawRecords.Response.Log)
+	}
+	records := make([]types.TimeLockRecord, 0)
+
+	if err = c.cdc.UnmarshalJSON(rawRecords.Response.GetValue(), &records); err != nil {
+		return nil, err
+	} else {
+		return records, nil
+	}
+
+}
+
+func (c *HTTP) GetTimelock(addr types.AccAddress, recordID int64) (*types.TimeLockRecord, error) {
+
+	params := types.QueryTimeLockParams{
+		Account: addr,
+		Id:      recordID,
+	}
+
+	bz, err := c.cdc.MarshalJSON(params)
+
+	if err != nil {
+		return nil, fmt.Errorf("incorrectly formatted request data %s", err.Error())
+	}
+
+	rawRecord, err := c.ABCIQuery(fmt.Sprintf("custom/%s/%s", TimeLockMsgRoute, "timelock"), bz)
+
+	if err != nil {
+		return nil, fmt.Errorf("error query %s", err.Error())
+	}
+	if rawRecord.Response.Code == TimeLockrcNotFoundErrorCode {
+		return nil, nil
+	}
+	var record types.TimeLockRecord
+
+	err = c.cdc.UnmarshalJSON(rawRecord.Response.GetValue(), &record)
+	if err != nil {
+		return nil, err
+	}
+
+	return &record, nil
+
 }
 
 func (c *HTTP) GetProposals(status types.ProposalStatus, numLatest int64) ([]types.Proposal, error) {
@@ -257,6 +376,9 @@ func (c *HTTP) GetProposals(status types.ProposalStatus, numLatest int64) ([]typ
 	if err != nil {
 		return nil, err
 	}
+	if !rawProposals.Response.IsOK(){
+		return nil, fmt.Errorf(rawProposals.Response.Log)
+	}
 	proposals := make([]types.Proposal, 0)
 
 	err = c.cdc.UnmarshalJSON(rawProposals.Response.GetValue(), &proposals)
@@ -271,14 +393,16 @@ func (c *HTTP) GetProposal(proposalId int64) (types.Proposal, error) {
 	if err != nil {
 		return nil, err
 	}
-	fmt.Println(string(bz))
-	rawProposals, err := c.ABCIQuery("custom/gov/proposal", bz)
+	rawProposal, err := c.ABCIQuery("custom/gov/proposal", bz)
 	if err != nil {
 		return nil, err
 	}
+	if !rawProposal.Response.IsOK(){
+		return nil, fmt.Errorf(rawProposal.Response.Log)
+	}
 	var proposal types.Proposal
 
-	err = c.cdc.UnmarshalJSON(rawProposals.Response.GetValue(), &proposal)
+	err = c.cdc.UnmarshalJSON(rawProposal.Response.GetValue(), &proposal)
 	return proposal, err
 }
 
@@ -299,4 +423,291 @@ func (c *HTTP) existsCC(symbol string) bool {
 		return false
 	}
 	return true
+}
+
+func (c *HTTP) GetSwapByID(swapID types.SwapBytes) (types.AtomicSwap, error) {
+	params := types.QuerySwapByID{
+		SwapID: swapID,
+	}
+	bz, err := c.cdc.MarshalJSON(params)
+	if err != nil {
+		return types.AtomicSwap{}, err
+	}
+
+	resp, err := c.ABCIQuery(fmt.Sprintf("custom/%s/%s", msg.AtomicSwapRoute, "swapid"), bz)
+	if err != nil {
+		return types.AtomicSwap{}, err
+	}
+	if !resp.Response.IsOK() {
+		return types.AtomicSwap{}, fmt.Errorf(resp.Response.Log)
+	}
+	if len(resp.Response.GetValue()) == 0 {
+		return types.AtomicSwap{}, fmt.Errorf("zero records")
+	}
+	var result types.AtomicSwap
+	err = c.cdc.UnmarshalJSON(resp.Response.GetValue(), &result)
+	if err != nil {
+		return types.AtomicSwap{}, err
+	}
+	return result, nil
+}
+
+func (c *HTTP) GetSwapByCreator(creatorAddr string, offset int64, limit int64) ([]types.SwapBytes, error) {
+	addr, err := types.AccAddressFromBech32(creatorAddr)
+	if err != nil {
+		return nil, err
+	}
+
+	params := types.QuerySwapByCreatorParams{
+		Creator: addr,
+		Limit:   limit,
+		Offset:  offset,
+	}
+
+	bz, err := c.cdc.MarshalJSON(params)
+	if err != nil {
+		return nil, err
+	}
+
+	resp, err := c.ABCIQuery(fmt.Sprintf("custom/%s/%s", msg.AtomicSwapRoute, "swapcreator"), bz)
+	if err != nil {
+		return nil, err
+	}
+	if !resp.Response.IsOK() {
+		return nil, fmt.Errorf(resp.Response.Log)
+	}
+	if len(resp.Response.GetValue()) == 0 {
+		return nil, fmt.Errorf("zero records")
+	}
+	var swapIDList []types.SwapBytes
+	err = c.cdc.UnmarshalJSON(resp.Response.GetValue(), &swapIDList)
+	if err != nil {
+		return nil, err
+	}
+	return swapIDList, nil
+}
+
+func (c *HTTP) GetSwapByRecipient(recipientAddr string, offset int64, limit int64) ([]types.SwapBytes, error) {
+	recipient, err := types.AccAddressFromBech32(recipientAddr)
+	if err != nil {
+		return nil, err
+	}
+	params := types.QuerySwapByRecipientParams{
+		Recipient: recipient,
+		Limit:     limit,
+		Offset:    offset,
+	}
+
+	bz, err := c.cdc.MarshalJSON(params)
+	if err != nil {
+		return nil, err
+	}
+
+	resp, err := c.ABCIQuery(fmt.Sprintf("custom/%s/%s", msg.AtomicSwapRoute, "swaprecipient"), bz)
+	if err != nil {
+		return nil, err
+	}
+	if !resp.Response.IsOK() {
+		return nil, fmt.Errorf(resp.Response.Log)
+	}
+	if len(resp.Response.GetValue()) == 0 {
+		return nil, fmt.Errorf("zero records")
+	}
+	var swapIDList []types.SwapBytes
+	err = c.cdc.UnmarshalJSON(resp.Response.GetValue(), &swapIDList)
+	if err != nil {
+		return nil, err
+	}
+	return swapIDList, nil
+}
+
+func (c *HTTP) SendToken(transfers []msg.Transfer, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error) {
+	if c.key == nil {
+		return nil, KeyMissingError
+	}
+	fromAddr := c.key.GetAddr()
+	fromCoins := types.Coins{}
+	for _, t := range transfers {
+		t.Coins = t.Coins.Sort()
+		fromCoins = fromCoins.Plus(t.Coins)
+	}
+	sendMsg := msg.CreateSendMsg(fromAddr, fromCoins, transfers)
+	return c.broadcast(sendMsg, syncType, options...)
+
+}
+
+func (c *HTTP) CreateOrder(baseAssetSymbol, quoteAssetSymbol string, op int8, price, quantity int64, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error) {
+	if c.key == nil {
+		return nil, KeyMissingError
+	}
+	if baseAssetSymbol == "" || quoteAssetSymbol == "" {
+		return nil, fmt.Errorf("BaseAssetSymbol or QuoteAssetSymbol is missing. ")
+	}
+	fromAddr := c.key.GetAddr()
+	newOrderMsg := msg.NewCreateOrderMsg(
+		fromAddr,
+		"",
+		op,
+		common.CombineSymbol(baseAssetSymbol, quoteAssetSymbol),
+		price,
+		quantity,
+	)
+	return c.broadcast(newOrderMsg, syncType, options...)
+}
+
+func (c *HTTP) HTLT(recipient types.AccAddress, recipientOtherChain, senderOtherChain string, randomNumberHash []byte, timestamp int64,
+	amount types.Coins, expectedIncome string, heightSpan int64, crossChain bool, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error) {
+	if c.key == nil {
+		return nil, KeyMissingError
+	}
+	fromAddr := c.key.GetAddr()
+	htltMsg := msg.NewHTLTMsg(
+		fromAddr,
+		recipient,
+		recipientOtherChain,
+		senderOtherChain,
+		randomNumberHash,
+		timestamp,
+		amount,
+		expectedIncome,
+		heightSpan,
+		crossChain,
+	)
+	return c.broadcast(htltMsg, syncType, options...)
+}
+
+func (c *HTTP) DepositHTLT(recipient types.AccAddress, swapID []byte, amount types.Coins,
+	syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error) {
+	if c.key == nil {
+		return nil, KeyMissingError
+	}
+	fromAddr := c.key.GetAddr()
+	depositHTLTMsg := msg.NewDepositHTLTMsg(
+		fromAddr,
+		swapID,
+		amount,
+	)
+	return c.broadcast(depositHTLTMsg, syncType, options...)
+}
+
+func (c *HTTP) ClaimHTLT(swapID []byte, randomNumber []byte, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error) {
+	if c.key == nil {
+		return nil, KeyMissingError
+	}
+	fromAddr := c.key.GetAddr()
+	claimHTLTMsg := msg.NewClaimHTLTMsg(
+		fromAddr,
+		swapID,
+		randomNumber,
+	)
+	return c.broadcast(claimHTLTMsg, syncType, options...)
+}
+
+func (c *HTTP) RefundHTLT(swapID []byte, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error) {
+	if c.key == nil {
+		return nil, KeyMissingError
+	}
+	fromAddr := c.key.GetAddr()
+	refundHTLTMsg := msg.NewRefundHTLTMsg(
+		fromAddr,
+		swapID,
+	)
+	return c.broadcast(refundHTLTMsg, syncType, options...)
+}
+
+func (c *HTTP) CancelOrder(baseAssetSymbol, quoteAssetSymbol, refId string, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error) {
+	if c.key == nil {
+		return nil, KeyMissingError
+	}
+	if baseAssetSymbol == "" || quoteAssetSymbol == "" {
+		return nil, fmt.Errorf("BaseAssetSymbol or QuoteAssetSymbol is missing. ")
+	}
+	if refId == "" {
+		return nil, fmt.Errorf("OrderId or Order RefId is missing. ")
+	}
+
+	fromAddr := c.key.GetAddr()
+
+	cancelOrderMsg := msg.NewCancelOrderMsg(fromAddr, common.CombineSymbol(baseAssetSymbol, quoteAssetSymbol), refId)
+	return c.broadcast(cancelOrderMsg, syncType, options...)
+}
+
+func (c *HTTP) broadcast(m msg.Msg, syncType SyncType, options ...tx.Option) (*core_types.ResultBroadcastTx, error) {
+	signBz, err := c.sign(m, options...)
+	if err != nil {
+		return nil, err
+	}
+	switch syncType {
+	case Async:
+		return c.BroadcastTxAsync(signBz)
+	case Sync:
+		return c.BroadcastTxSync(signBz)
+	case Commit:
+		commitRes, err := c.BroadcastTxCommit(signBz)
+		if err != nil {
+			return nil, err
+		}
+		if commitRes.CheckTx.IsErr() {
+			return &core_types.ResultBroadcastTx{
+				Code: commitRes.CheckTx.Code,
+				Log:  commitRes.CheckTx.Log,
+				Hash: commitRes.Hash,
+				Data: commitRes.CheckTx.Data,
+			}, nil
+		}
+		return &core_types.ResultBroadcastTx{
+			Code: commitRes.DeliverTx.Code,
+			Log:  commitRes.DeliverTx.Log,
+			Hash: commitRes.Hash,
+			Data: commitRes.DeliverTx.Data,
+		}, nil
+	default:
+		return nil, fmt.Errorf("unknown synctype")
+	}
+}
+
+func (c *HTTP) sign(m msg.Msg, options ...tx.Option) ([]byte, error) {
+	if c.key == nil {
+		return nil, fmt.Errorf("keymanager is missing, use SetKeyManager to set key")
+	}
+	// prepare message to sign
+	chainID := gtypes.ProdChainID
+	if types.Network != types.ProdNetwork {
+		chainID = gtypes.TestnetChainID
+	}
+	signMsg := &tx.StdSignMsg{
+		ChainID:       chainID,
+		AccountNumber: -1,
+		Sequence:      -1,
+		Memo:          "",
+		Msgs:          []msg.Msg{m},
+		Source:        tx.Source,
+	}
+
+	for _, op := range options {
+		signMsg = op(signMsg)
+	}
+
+	if signMsg.Sequence == -1 || signMsg.AccountNumber == -1 {
+		fromAddr := c.key.GetAddr()
+		acc, err := c.GetAccount(fromAddr)
+		if err != nil {
+			return nil, err
+		}
+		signMsg.Sequence = acc.GetSequence()
+		signMsg.AccountNumber = acc.GetAccountNumber()
+	}
+
+	// special logic for createOrder, to save account query
+	if orderMsg, ok := m.(msg.CreateOrderMsg); ok {
+		orderMsg.ID = msg.GenerateOrderID(signMsg.Sequence+1, c.key.GetAddr())
+		signMsg.Msgs[0] = orderMsg
+	}
+
+	for _, m := range signMsg.Msgs {
+		if err := m.ValidateBasic(); err != nil {
+			return nil, err
+		}
+	}
+	return c.key.Sign(*signMsg)
 }